completed wrapped decoder. issue with saving values to db with repos--ids are created linearly, while foreign keys may not yet exist.

+++ dat/src/main/java/com/github/poeatlas/cli/dat/domain/AtlasNode.java
@@ -34,10 +34,10 @@ public class AtlasNode {
   private long worldAreasKey;
 
   @Column(nullable = false)
-  private double posX;
+  private float posX;
 
   @Column(nullable = false)
-  private double posY;
+  private float posY;
 
   @Transient
   private int unknown4;
